Move FeedErrorViewModel to separate file
