+++ EssentialFeed/EssentialFeed/Feed Presentation/FeedPresenter.swift
@@ -15,19 +15,6 @@ public protocol FeedLoadingView {
   func display(_ viewModel: FeedLoadingViewModel)
 }
 
-public struct FeedErrorViewModel {
-  
-  public let message: String?
-  
-  static var noError: FeedErrorViewModel {
-    return FeedErrorViewModel(message: nil)
-  }
-  
-  static func error(message: String) -> FeedErrorViewModel {
-    return FeedErrorViewModel(message: message)
-  }
-}
-
 public protocol FeedErrorView {
   func display(_ viewModel: FeedErrorViewModel)
 }
